% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{get_humicon}
\alias{get_humicon}
\title{Get OCHA Humanitarian icons v02 unicode}
\usage{
get_humicon(name)
}
\arguments{
\item{name}{character name of the icon see humicons_data}
}
\description{
Get you the unicode to use in base graphic or ggplot2
}
\examples{
get_humicon("Urban")

}
