% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.R
\name{hi-alias}
\alias{hi-alias}
\title{Humicons alias}
\description{
Humicons alias
}
\keyword{internal} \alias{hi_Abduction_kidnapping}
\keyword{internal} \alias{hi_About}
\keyword{internal} \alias{hi_Add_document}
\keyword{internal} \alias{hi_Add}
\keyword{internal} \alias{hi_Advocacy}
\keyword{internal} \alias{hi_Affected_population}
\keyword{internal} \alias{hi_Agriculture}
\keyword{internal} \alias{hi_Airport_affected}
\keyword{internal} \alias{hi_Airport_destroyed}
\keyword{internal} \alias{hi_Airport_military}
\keyword{internal} \alias{hi_Airport_not_affected}
\keyword{internal} \alias{hi_Airport}
\keyword{internal} \alias{hi_Alert}
\keyword{internal} \alias{hi_Analysis}
\keyword{internal} \alias{hi_Arrest_detention}
\keyword{internal} \alias{hi_Assault}
\keyword{internal} \alias{hi_Assembly_point}
\keyword{internal} \alias{hi_Assessment}
\keyword{internal} \alias{hi_Attack}
\keyword{internal} \alias{hi_Blanket}
\keyword{internal} \alias{hi_Blog}
\keyword{internal} \alias{hi_Boat}
\keyword{internal} \alias{hi_Bookmark}
\keyword{internal} \alias{hi_Border_crossing}
\keyword{internal} \alias{hi_Borehole}
\keyword{internal} \alias{hi_Bottled_water}
\keyword{internal} \alias{hi_Bridge_affected}
\keyword{internal} \alias{hi_Bridge_destroyed}
\keyword{internal} \alias{hi_Bridge_not_affected}
\keyword{internal} \alias{hi_Bridge}
\keyword{internal} \alias{hi_Bucket}
\keyword{internal} \alias{hi_Buddhist_temple}
\keyword{internal} \alias{hi_Building_facility_affected}
\keyword{internal} \alias{hi_Building_facility_destroyed}
\keyword{internal} \alias{hi_Building_facility_not_affected}
\keyword{internal} \alias{hi_Building}
\keyword{internal} \alias{hi_Bus}
\keyword{internal} \alias{hi_Calendar}
\keyword{internal} \alias{hi_Camp_Coordination_and_Camp_Management}
\keyword{internal} \alias{hi_Car}
\keyword{internal} \alias{hi_Carjacking}
\keyword{internal} \alias{hi_Cash_transfer}
\keyword{internal} \alias{hi_Cell_tower}
\keyword{internal} \alias{hi_Chart}
\keyword{internal} \alias{hi_Chat}
\keyword{internal} \alias{hi_Checkpoint}
\keyword{internal} \alias{hi_Child_combatant}
\keyword{internal} \alias{hi_Children}
\keyword{internal} \alias{hi_Church}
\keyword{internal} \alias{hi_Civil_military_coordination}
\keyword{internal} \alias{hi_Clinic}
\keyword{internal} \alias{hi_Clothing}
\keyword{internal} \alias{hi_Cold_wave}
\keyword{internal} \alias{hi_Communal_latrine}
\keyword{internal} \alias{hi_Community_building}
\keyword{internal} \alias{hi_Community_engagement}
\keyword{internal} \alias{hi_Computer}
\keyword{internal} \alias{hi_Conflict}
\keyword{internal} \alias{hi_Coordinated_assessement}
\keyword{internal} \alias{hi_Coordination}
\keyword{internal} \alias{hi_Copy}
\keyword{internal} \alias{hi_Cyclone}
\keyword{internal} \alias{hi_Damaged_Affected}
\keyword{internal} \alias{hi_Dangerous_area}
\keyword{internal} \alias{hi_Data}
\keyword{internal} \alias{hi_Dead}
\keyword{internal} \alias{hi_Debris_management}
\keyword{internal} \alias{hi_Deployment}
\keyword{internal} \alias{hi_Destroyed}
\keyword{internal} \alias{hi_Detergent}
\keyword{internal} \alias{hi_Diplomatic_mission}
\keyword{internal} \alias{hi_Distribution_site}
\keyword{internal} \alias{hi_Document}
\keyword{internal} \alias{hi_Download}
\keyword{internal} \alias{hi_Drought}
\keyword{internal} \alias{hi_Drowned}
\keyword{internal} \alias{hi_E_mail}
\keyword{internal} \alias{hi_Early_Recovery}
\keyword{internal} \alias{hi_Earthmound}
\keyword{internal} \alias{hi_Earthquake}
\keyword{internal} \alias{hi_Education}
\keyword{internal} \alias{hi_Elderly}
\keyword{internal} \alias{hi_Emergency_Telecommunications}
\keyword{internal} \alias{hi_Environment}
\keyword{internal} \alias{hi_Epidemic}
\keyword{internal} \alias{hi_Exit_Cancel}
\keyword{internal} \alias{hi_Favourite}
\keyword{internal} \alias{hi_Fax}
\keyword{internal} \alias{hi_Ferry}
\keyword{internal} \alias{hi_Film}
\keyword{internal} \alias{hi_Financing}
\keyword{internal} \alias{hi_Fire}
\keyword{internal} \alias{hi_Fishery}
\keyword{internal} \alias{hi_Flash_flood}
\keyword{internal} \alias{hi_Flood}
\keyword{internal} \alias{hi_Flour}
\keyword{internal} \alias{hi_Folder}
\keyword{internal} \alias{hi_Food_Security}
\keyword{internal} \alias{hi_Food_warehouse}
\keyword{internal} \alias{hi_Food}
\keyword{internal} \alias{hi_Forced_entry}
\keyword{internal} \alias{hi_Forced_recruitment}
\keyword{internal} \alias{hi_Fund}
\keyword{internal} \alias{hi_Gap_analysis}
\keyword{internal} \alias{hi_Gas_station}
\keyword{internal} \alias{hi_Go}
\keyword{internal} \alias{hi_Government_office}
\keyword{internal} \alias{hi_Group}
\keyword{internal} \alias{hi_Harassment_intimidation}
\keyword{internal} \alias{hi_Health_facility}
\keyword{internal} \alias{hi_Health_post}
\keyword{internal} \alias{hi_Health}
\keyword{internal} \alias{hi_Heatwave}
\keyword{internal} \alias{hi_Heavy_rain}
\keyword{internal} \alias{hi_Helicopter}
\keyword{internal} \alias{hi_Helipad}
\keyword{internal} \alias{hi_Help}
\keyword{internal} \alias{hi_Hidden}
\keyword{internal} \alias{hi_Hindu_temple}
\keyword{internal} \alias{hi_Hospital}
\keyword{internal} \alias{hi_Hotel}
\keyword{internal} \alias{hi_House_affected}
\keyword{internal} \alias{hi_House_burned}
\keyword{internal} \alias{hi_House_destroyed}
\keyword{internal} \alias{hi_House_not_affected}
\keyword{internal} \alias{hi_House}
\keyword{internal} \alias{hi_Humanitarian_access}
\keyword{internal} \alias{hi_Humanitarian_programme_cycle}
\keyword{internal} \alias{hi_IDP_refugee_camp}
\keyword{internal} \alias{hi_Infant_formula}
\keyword{internal} \alias{hi_Infant}
\keyword{internal} \alias{hi_Information_management}
\keyword{internal} \alias{hi_Information_technology}
\keyword{internal} \alias{hi_Infrastructure}
\keyword{internal} \alias{hi_Injured}
\keyword{internal} \alias{hi_Innovation}
\keyword{internal} \alias{hi_Insect_infestation}
\keyword{internal} \alias{hi_Internally_displaced}
\keyword{internal} \alias{hi_Internet}
\keyword{internal} \alias{hi_Kitchen_set}
\keyword{internal} \alias{hi_Landslide_mudslide}
\keyword{internal} \alias{hi_Laptop}
\keyword{internal} \alias{hi_Latrine_cabin}
\keyword{internal} \alias{hi_Leadership}
\keyword{internal} \alias{hi_Learning}
\keyword{internal} \alias{hi_Livelihood}
\keyword{internal} \alias{hi_Livestock}
\keyword{internal} \alias{hi_Location}
\keyword{internal} \alias{hi_Locust_infestation}
\keyword{internal} \alias{hi_Logistics}
\keyword{internal} \alias{hi_Map}
\keyword{internal} \alias{hi_Mattress}
\keyword{internal} \alias{hi_Medical_supply}
\keyword{internal} \alias{hi_Medicine}
\keyword{internal} \alias{hi_Meeting}
\keyword{internal} \alias{hi_Menu}
\keyword{internal} \alias{hi_Military_gate}
\keyword{internal} \alias{hi_Mine}
\keyword{internal} \alias{hi_Missing}
\keyword{internal} \alias{hi_Mobile_clinic}
\keyword{internal} \alias{hi_Mobile_phone}
\keyword{internal} \alias{hi_Monitor}
\keyword{internal} \alias{hi_Monitoring}
\keyword{internal} \alias{hi_More_options}
\keyword{internal} \alias{hi_Mosque}
\keyword{internal} \alias{hi_Mosquito_net}
\keyword{internal} \alias{hi_Multi_cluster_sector}
\keyword{internal} \alias{hi_Murder}
\keyword{internal} \alias{hi_National_army}
\keyword{internal} \alias{hi_Needs_assessment}
\keyword{internal} \alias{hi_Next_item}
\keyword{internal} \alias{hi_NGO_office}
\keyword{internal} \alias{hi_Non_food_items}
\keyword{internal} \alias{hi_Not_affected}
\keyword{internal} \alias{hi_Notification}
\keyword{internal} \alias{hi_Nutrition}
\keyword{internal} \alias{hi_Observation_tower}
\keyword{internal} \alias{hi_Oil}
\keyword{internal} \alias{hi_Out_of_platform}
\keyword{internal} \alias{hi_Partnership}
\keyword{internal} \alias{hi_Pause}
\keyword{internal} \alias{hi_Peacekeeping_force}
\keyword{internal} \alias{hi_People_with_physical_impairments}
\keyword{internal} \alias{hi_Permanent_camp}
\keyword{internal} \alias{hi_Person_1}
\keyword{internal} \alias{hi_Person_2}
\keyword{internal} \alias{hi_Photo}
\keyword{internal} \alias{hi_Physical_closure}
\keyword{internal} \alias{hi_Plastic_sheeting}
\keyword{internal} \alias{hi_Police_station}
\keyword{internal} \alias{hi_Policy}
\keyword{internal} \alias{hi_Population_growth}
\keyword{internal} \alias{hi_Population_return}
\keyword{internal} \alias{hi_Port_affected}
\keyword{internal} \alias{hi_Port_destroyed}
\keyword{internal} \alias{hi_Port_not_affected}
\keyword{internal} \alias{hi_Port}
\keyword{internal} \alias{hi_Potable_water_source}
\keyword{internal} \alias{hi_Potable_water}
\keyword{internal} \alias{hi_Pregnant}
\keyword{internal} \alias{hi_Preparedness}
\keyword{internal} \alias{hi_Previous_item}
\keyword{internal} \alias{hi_Print}
\keyword{internal} \alias{hi_Protection}
\keyword{internal} \alias{hi_Public_information}
\keyword{internal} \alias{hi_Radio}
\keyword{internal} \alias{hi_Rebel}
\keyword{internal} \alias{hi_Reconstruction}
\keyword{internal} \alias{hi_Refugee}
\keyword{internal} \alias{hi_Registration}
\keyword{internal} \alias{hi_Relief_goods}
\keyword{internal} \alias{hi_Remove_document}
\keyword{internal} \alias{hi_Remove}
\keyword{internal} \alias{hi_Report}
\keyword{internal} \alias{hi_Reporting}
\keyword{internal} \alias{hi_Response}
\keyword{internal} \alias{hi_Return}
\keyword{internal} \alias{hi_Rice}
\keyword{internal} \alias{hi_Road_affected}
\keyword{internal} \alias{hi_Road_barrier}
\keyword{internal} \alias{hi_Road_destroyed}
\keyword{internal} \alias{hi_Road_not_affected}
\keyword{internal} \alias{hi_Road}
\keyword{internal} \alias{hi_Roadblock}
\keyword{internal} \alias{hi_Robbery}
\keyword{internal} \alias{hi_Rule_of_law_and_justice}
\keyword{internal} \alias{hi_Rural_exodus}
\keyword{internal} \alias{hi_Rural}
\keyword{internal} \alias{hi_Safety_and_security}
\keyword{internal} \alias{hi_Salt}
\keyword{internal} \alias{hi_Sanitation}
\keyword{internal} \alias{hi_Satellite_dish}
\keyword{internal} \alias{hi_Save}
\keyword{internal} \alias{hi_Scale_down_operation}
\keyword{internal} \alias{hi_Scale_up_operation}
\keyword{internal} \alias{hi_School_affected}
\keyword{internal} \alias{hi_School_destroyed}
\keyword{internal} \alias{hi_School_not_affected}
\keyword{internal} \alias{hi_School}
\keyword{internal} \alias{hi_Search}
\keyword{internal} \alias{hi_See}
\keyword{internal} \alias{hi_Selected}
\keyword{internal} \alias{hi_Services_and_tools}
\keyword{internal} \alias{hi_Settings}
\keyword{internal} \alias{hi_Sexual_violence}
\keyword{internal} \alias{hi_Share}
\keyword{internal} \alias{hi_Shelter}
\keyword{internal} \alias{hi_Ship}
\keyword{internal} \alias{hi_Shower}
\keyword{internal} \alias{hi_Smartphone}
\keyword{internal} \alias{hi_Snow_avalanche}
\keyword{internal} \alias{hi_Snowfall}
\keyword{internal} \alias{hi_Soap}
\keyword{internal} \alias{hi_Solid_waste}
\keyword{internal} \alias{hi_Spontaneous_site}
\keyword{internal} \alias{hi_Spring_water}
\keyword{internal} \alias{hi_Staff_management}
\keyword{internal} \alias{hi_Stop}
\keyword{internal} \alias{hi_Storm_surge}
\keyword{internal} \alias{hi_Storm}
\keyword{internal} \alias{hi_Stove}
\keyword{internal} \alias{hi_Submersible_pump}
\keyword{internal} \alias{hi_Sugar}
\keyword{internal} \alias{hi_Table}
\keyword{internal} \alias{hi_Tarpaulin}
\keyword{internal} \alias{hi_Technological_disaster}
\keyword{internal} \alias{hi_Temporary_camp}
\keyword{internal} \alias{hi_Tent}
\keyword{internal} \alias{hi_Toilet}
\keyword{internal} \alias{hi_Tornado}
\keyword{internal} \alias{hi_Trade_and_market}
\keyword{internal} \alias{hi_Train}
\keyword{internal} \alias{hi_Transition_site}
\keyword{internal} \alias{hi_Trending}
\keyword{internal} \alias{hi_Truck}
\keyword{internal} \alias{hi_Tsunami}
\keyword{internal} \alias{hi_Tunnel}
\keyword{internal} \alias{hi_UN_compound_office}
\keyword{internal} \alias{hi_UN_vehicle}
\keyword{internal} \alias{hi_University}
\keyword{internal} \alias{hi_Upload}
\keyword{internal} \alias{hi_Urban_rural}
\keyword{internal} \alias{hi_Urban}
\keyword{internal} \alias{hi_User}
\keyword{internal} \alias{hi_Users}
\keyword{internal} \alias{hi_Vaccine}
\keyword{internal} \alias{hi_Video}
\keyword{internal} \alias{hi_Violent_wind}
\keyword{internal} \alias{hi_Volcano}
\keyword{internal} \alias{hi_Walkie_talkie}
\keyword{internal} \alias{hi_Warning_Error}
\keyword{internal} \alias{hi_Water_Sanitation_and_Hygiene}
\keyword{internal} \alias{hi_Water_source}
\keyword{internal} \alias{hi_Water_trucking}
\keyword{internal} \alias{hi_ZIP_compressed}
