% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\docType{package}
\name{humicon-package}
\alias{humicon}
\alias{humicon-package}
\title{humicon: Humanitarian Icon web-fonts for RMarkdown}
\description{
This package allows easy insertion of humanitarian icons into RMarkdown.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/dickoa/humicon}
  \item Report bugs at \url{https://gitlab.com/dickoa/humicon/issues}
}

}
\author{
\strong{Maintainer}: Ahmadou Dicko \email{mail@ahmadoudicko.com}

}
\keyword{internal}
