% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humicon.R
\name{humicon}
\alias{humicon}
\title{Insert icon from OCHA humanitarian icons v2.0.0 in a Shiny app}
\usage{
humicon(name, size = 1, fixed_width = FALSE, animate = "still",
  rotate = 0, flip = "none", border = FALSE, pull = NULL,
  color = NULL, colour = color, other = NULL, class = NULL)
}
\arguments{
\item{name}{A string indicating the icon name.}

\item{size}{Size of the icon relative to font size. Options are 1, lg (33%
increase), 2, 3, 4, or 5.}

\item{fixed_width}{If TRUE, the icon is set to a fixed width}

\item{animate}{'still', 'spin', or 'pulse'.}

\item{rotate}{Rotate degree: 0, 90, 180, or 270.}

\item{flip}{'none', 'horizontal', 'vertical'.}

\item{border}{If TRUE, draws a border around the icon.}

\item{pull}{Pulls icon to either 'left' or 'right' and wraps proceeding text
around it.}

\item{color, colour}{Hex code for a colour to be given to the icon}

\item{other}{Character vector of other parameters directly added to the icon classes}
}
\value{
An Shiny icon element
}
\description{
Create an icon for use within a page in a Shiny app. Icons can appear on their own, inside
of a button, or as an icon for a \code{\link{tabPanel}} within a
\code{\link{navbarPage}}.
}
\examples{
# add an icon to a submit button
submitButton("Update View", icon = icon("Fund"))

navbarPage("App Title",
  tabPanel("Plot", icon = humicon("Food-Security")),
  tabPanel("Summary", icon = humicon("Agriculture")),
  tabPanel("Table", icon = humicon("Livestock"))
)
}
