% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humicon.R
\name{humicon_chr}
\alias{humicon_chr}
\title{Get OCHA Humanitarian icons v02 unicode in character}
\usage{
humicon_chr(names)
}
\arguments{
\item{names}{character name of the icon see humicons_data}
}
\description{
Get you the unicode to use in base graphic or ggplot2
}
\examples{
humicon_chr(c("Urban", "Car"))

}
