% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humicons-package.R
\docType{package}
\name{humicons-package}
\alias{humicons}
\alias{humicons-package}
\title{humicons: OCHA Humanitarian Icons v02 web-font for RMarkdown}
\description{
This package allows easy insertion of the new set of humanitarian icons into RMarkdown.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/dickoa/humicons}
  \item Report bugs at \url{https://gitlab.com/dickoa/humicons/issues}
}

}
\author{
\strong{Maintainer}: Ahmadou Dicko \email{mail@ahmadoudicko.com}

}
\keyword{internal}
