% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{humicons}
\alias{humicons}
\title{Get OCHA Humanitarian icons v02 unicode}
\usage{
humicons(names)
}
\arguments{
\item{names}{character name of the icon see humicons_data}
}
\description{
Get you the unicode to use in base graphic or ggplot2
}
\examples{
humicons(c("Urban", "Car"))

}
