% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humicons_data}
\alias{humicons_data}
\title{OCHA Humanitarian fonts v02 data in tibble format}
\format{A tibble with 295 rows and 2 variables:
\describe{
\item{name}{name of the icon}
\item{code}{code of the icon}
}}
\source{
\url{https://thenounproject.com/ochavisual/collection/ocha-humanitarian-icons-v02/}
}
\usage{
humicons_data
}
\description{
A dataset containing the name and code for all 295 humanitarian icons
}
\keyword{datasets}
